Scripts and Docker file to work on the repository
I don't want to install all the dependencies, or add external tools like
perlenv and perlbrew to isolate the dependencies.

The Docker file creates a basic container with all the required
dependencies, and the scripts make it easy to run the tests inside the
container.

A Makefile is also added to build the image, run the tests, and so on.